refactor(p4-3/4): split the monolith along its seams — reads, held-queue, assignment
prospectService drops from 3,089 to ~1,700 lines (capture core, update/
delete, composition). Three sub-factories receive the parent's exact (d, m)
so every test override keeps flowing through makeProspectService unchanged:

- prospectReadService.makeProspectReads: detail/stats/list/held-list/
  activities/view-tracking projections.
- prospectHeldQueueService.makeHeldQueueOps: orphan + held-lead lifecycle,
  taking the shared idempotency ops and assignProspect (the single
  cross-seam dependency, verified by static call-map before the cut).
- prospectAssignmentService.makeAssignmentOps: single + bulk assignment.

Public API byte-compatible: makeProspectService's return shape and every
named export unchanged; zero test modifications.
