+++ backend/src/services/prospectService.js
@@ -1,5 +1,5 @@
-import { randomUUID, createHash } from 'crypto';
-import { Op, Transaction, fn as sqlFn, col as sqlCol, where as sqlWhere } from 'sequelize';
+import { createHash } from 'crypto';
+import { Transaction } from 'sequelize';
 import {
   Prospect,
   User,
@@ -14,18 +14,18 @@ import {
 } from '../models/index.js';
 import { resolveLeadRouting, getSystemAgentId, resolveLeadAssignment } from './systemAgent.js';
 import { makeIdempotencyOps } from './idempotencyProtocol.js';
+import { makeProspectReads } from './prospectReadService.js';
+import { makeHeldQueueOps } from './prospectHeldQueueService.js';
+import { makeAssignmentOps } from './prospectAssignmentService.js';
 import {
-  UUID_RE, VALID_LEAD_STATUSES, VALID_ASSIGNMENT_FILTERS, VALID_LEAD_SOURCES,
-  parseEnumFilter, parseProspectSort, RELEASABLE_HOLD_REASONS, PROSPECT_UPDATE_FIELDS,
+  UUID_RE, PROSPECT_UPDATE_FIELDS,
 } from './prospectShared.js';
 import { deductLeadCredit, chargeLeadCredit, deductExternalLeadBalance } from './leadCredits.js';
 import { decideAssignment } from './leadQuota.js';
 import { dncEnforcement, formatDncNumber, checkAndRecord as dncCheckAndRecord } from './dncService.js';
 import { gateHeldDncLead } from './dncGate.js';
-import { SCREENING_REASONS } from './screeningConstants.js';
 import { hasValidExternalConsent, buildExternalConsentEvidence } from './externalConsent.js';
 import { buildDncConsentEvidence } from './dncConsent.js';
-import { repeatSignupDetail, repeatSignupCounts } from './repeatSignup.js';
 import { buildProspectWhere } from '../middleware/prospectScope.js';
 import { AppError } from '../middleware/errorHandler.js';
 import { dispatchEvent, persistEventDeliveries, flushDeliveries, hasDeliverableSubscriber } from './webhookService.js';
@@ -74,20 +74,15 @@ export function registerLeadCapturedHook(fn) {
   _leadCapturedHook = typeof fn === 'function' ? fn : null;
 }
 import { logger } from '../utils/logger.js';
-import { signupActivityDescription, signupSourceLabel } from '../utils/sourceLabel.js';
+import { signupActivityDescription } from '../utils/sourceLabel.js';
 import {
   normalizePhone,
   buildLeadCreatedPayload,
   buildLeadHeldPayload,
   buildLeadDeletedPayload,
-  buildLeadAssignedPayload,
-  buildLeadUnassignedPayload,
-  buildHeldLeadEnrichment,
   destinationForAgent,
   externalIdForDestination,
-  withBatchContext,
 } from './prospectHelpers.js';
-import { fetchLeadActivitiesFromSupabase, mergeProspectTimeline } from './webLeadTimelineService.js';
 import { scoreQuiz } from './quizScoringService.js';
 import { readLegacyViewSafe } from '../utils/designConfigV2Clamp.js';
 
@@ -155,6 +150,12 @@ export function makeProspectService(overrides = {}) {
   const d = { ...defaultDeps, ...overrides };
   const m = { ...defaultDeps.models, ...(overrides.models || {}) };
   const idem = makeIdempotencyOps(m.IdempotencyKey);
+  const reads = makeProspectReads({ d, m });
+  const { getProspect, getProspectStats, listProspects, trackProspectView, listHeldProspects, getProspectActivities } = reads;
+  const assigns = makeAssignmentOps({ d, m });
+  const { assignProspect, bulkAssignProspects } = assigns;
+  const held = makeHeldQueueOps({ d, m, idem, assignProspect });
+  const { orphanSystemAgentId, listDispatchableOrphans, releaseHeldProspect, reassignProspectExternal, returnProspectToHeld, bulkReturnProspectsToHeld } = held;
 
   /**
    * Create a new prospect (lead capture).
@@ -1377,122 +1378,6 @@ export function makeProspectService(overrides = {}) {
   }
 
   /**
-   * Get a single prospect by ID, scoped to user access.
-   *
-   * `include: 'profile'` (Lead Profile page, plan §4) is ADMIN-ONLY and
-   * opt-in: non-admins and callers that don't ask get a byte-identical
-   * payload to before the option existed — the flag simply never reaches the
-   * enrichment branch for them.
-   */
-  async function getProspect(id, user, { include = '' } = {}) {
-    const wantProfile = user?.role === 'admin'
-      && String(include).split(',').map((s) => s.trim()).includes('profile');
-    const scopeFilter = await d.buildProspectWhere(user);
-    const whereConditions = { id, ...scopeFilter };
-
-    const prospect = await m.Prospect.findOne({
-      where: whereConditions,
-      include: [
-        {
-          association: 'assignedAgent',
-          attributes: ['id', 'firstName', 'lastName', 'email', 'phone'],
-        },
-        {
-          association: 'campaign',
-          attributes: ['id', 'name', 'type', 'status', 'description'],
-        },
-        {
-          association: 'qrTag',
-          attributes: ['id', 'name', 'type', 'location'],
-        },
-        // Named external buyer — profile mode only (the association exists but
-        // was never loaded; keep the classic payload byte-identical). Name and
-        // agency ONLY: never phone/email/balance on an admin lead view.
-        ...(wantProfile ? [{
-          association: 'externalAgent',
-          attributes: ['id', 'fullName', 'agency'],
-        }] : []),
-        {
-          association: 'activities',
-          // Newest-first, so the Activity Timeline UI (which renders this array
-          // top-to-bottom and pins a "Start of History" marker at the bottom) leads
-          // with the most recent event and ends with the genesis 'created' event.
-          //
-          // `separate: true` is REQUIRED: an `order` inside a normal (JOINed) hasMany
-          // include is silently ignored by Sequelize v6 — only the separate (own-query)
-          // loader honours include-local ordering. `prospectId` must stay in
-          // `attributes` because that loader maps children back to the parent by the
-          // FK; omitting it returns an empty activities array. `id DESC` is a
-          // deterministic tiebreaker for same-millisecond `createdAt` values.
-          separate: true,
-          attributes: ['id', 'prospectId', 'type', 'description', 'metadata', 'createdAt'],
-          order: [['createdAt', 'DESC'], ['id', 'DESC']],
-        },
-      ],
-    });
-
-    if (!prospect) {
-      throw new d.AppError('Prospect not found or access denied', 404);
-    }
-
-    // Admin-only: cross-campaign repeat-signup visibility (flag, not block).
-    // Omitted entirely for non-admins (this endpoint is shared with the agent
-    // detail modal). Resilient — a failed enrichment never breaks the view.
-    // Admin-only enrichments. The unified `timeline` merges the agent-engagement half (Supabase
-    // lead_activities) with this prospect's ProspectActivity so the web Activity Timeline matches
-    // the mktr-leads app. ADMIN-GATED on purpose: the engagement is the external buyer's private
-    // notes, which must not surface to a non-admin web user (agents can open the same detail modal).
-    // Gated additionally on the export-EF URL (deploy-inert until set). Run in PARALLEL with the
-    // repeat-signup lookup so neither blocks the other; both are resilient (a miss never breaks the
-    // view, and the timeline degrades to ProspectActivity-only on a Supabase miss).
-    if (user?.role === 'admin') {
-      const wantTimeline = !!process.env.SUPABASE_LEAD_ACTIVITIES_URL;
-      const [repeatSignup, timelineFetch, consumerJourney, session, lyfeDelivery, signupProfile] = await Promise.all([
-        repeatSignupDetail(d.sequelize, { phone: prospect.phone, email: prospect.email }).catch(() => null),
-        wantTimeline
-          ? fetchLeadActivitiesFromSupabase(prospect.id).catch(() => ({ rows: [], ok: false }))
-          : Promise.resolve(null),
-        // Consumer spine: the person's cross-campaign journey (signups +
-        // rewards) for the admin drawer's Person card. Resilient like its
-        // siblings — a miss never breaks the detail view. Profile mode
-        // (Lead Profile page) enriches the same journey in place: per-signup
-        // draw standing + scoped consent + reward diagnostics, entitlement
-        // presentation extras, suppressions, broadcast history.
-        prospect.consumerId
-          ? d.getConsumerJourney(prospect.consumerId, wantProfile ? { includeRaw: true } : undefined)
-            .then((j) => (wantProfile && j ? d.enrichJourneyProfile(j) : j))
-            .catch(() => null)
-          : Promise.resolve(null),
-        // Profile-only extras (each resilient, each bounded — plan §4 B6/B7).
-        wantProfile
-          ? d.getSessionContext(prospect.sessionId).catch(() => null)
-          : Promise.resolve(null),
-        wantProfile
-          ? d.getLyfeDelivery(prospect.id).catch(() => null)
-          : Promise.resolve(null),
-        // B4: a consumer-less lead (Retell, pre-spine, unverified) still gets
-        // its OWN campaign's draw/reward standing.
-        wantProfile && !prospect.consumerId
-          ? d.getSignupProfile(prospect).catch(() => null)
-          : Promise.resolve(null),
-      ]);
-      if (repeatSignup) prospect.setDataValue('repeatSignup', repeatSignup);
-      if (consumerJourney) prospect.setDataValue('consumer', consumerJourney);
-      if (timelineFetch) {
-        prospect.setDataValue(
-          'timeline',
-          mergeProspectTimeline(prospect.activities || [], timelineFetch.rows, { ok: timelineFetch.ok }),
-        );
-      }
-      if (session) prospect.setDataValue('session', session);
-      if (lyfeDelivery) prospect.setDataValue('lyfeDelivery', lyfeDelivery);
-      if (signupProfile) prospect.setDataValue('signupProfile', signupProfile);
-    }
-
-    return prospect;
-  }
-
-  /**
    * Update a prospect. Handles status-change-to-won commission logic.
    */
   async function updateProspect(id, body, user) {
@@ -1734,669 +1619,6 @@ export function makeProspectService(overrides = {}) {
   }
 
   /**
-   * Assign a single prospect to an agent. Returns { prospect, agent } for email side-effect.
-   * `opts.batch` ({ id, size }, pre-validated) rides into the delivery webhooks so the
-   * receiving app can coalesce a bulk fan-out's pushes into one summary.
-   */
-  async function assignProspect(prospectId, agentId, user, opts = {}) {
-    const { batch = null } = opts;
-    const prospect = await m.Prospect.findByPk(prospectId);
-    if (!prospect) {
-      throw new d.AppError('Prospect not found', 404);
-    }
-
-    const previousAgentId = prospect.assignedAgentId;
-
-    // ── Unassign ──
-    if (!agentId) {
-      await prospect.update({ assignedAgentId: null });
-
-      await m.ProspectActivity.create({
-        prospectId: prospect.id,
-        type: 'assigned',
-        actorUserId: user?.id || null,
-        description: 'Unassigned from agent',
-        metadata: { previousAgentId },
-      });
-
-      // Fire lead.unassigned webhook — but NOT for a HELD (quarantined) lead: Lyfe never
-      // received a lead.created for it (the create webhook was suppressed), so an
-      // unassigned event would reference a lead Lyfe does not know about.
-      if (!prospect.quarantinedAt) {
-        // Destination + external id come from the PREVIOUS agent (the assignment
-        // is already null). Sourceless previous agent -> null -> default-denied.
-        let prevDestination = null;
-        let previousAgentExternalId = null;
-        if (previousAgentId) {
-          const prevAgent = await m.User.findByPk(previousAgentId, {
-            attributes: ['id', 'lyfeId', 'mktrLeadsId'],
-          });
-          prevDestination = destinationForAgent(prevAgent);
-          previousAgentExternalId = externalIdForDestination(prevAgent, prevDestination);
-        }
-        d.dispatchEvent('lead.unassigned', () => buildLeadUnassignedPayload(prospect, previousAgentExternalId), {
-          destination: prevDestination,
-        });
-      }
-
-      return { prospect, agent: null, prospectWithCampaign: prospect };
-    }
-
-    // ── Assign ──
-    const agent = await m.User.findOne({
-      where: { id: agentId, role: 'agent', isActive: true },
-    });
-
-    if (!agent) {
-      throw new d.AppError('Invalid or inactive agent', 400);
-    }
-
-    // An external hold (no funded MKTR Leads buyer) must NEVER be manually released to an
-    // internal agent / Lyfe — it was captured for the external buyer pool and can only be
-    // delivered via the external channel (or a dedicated conversion flow, not built yet).
-    // The auto release-sweep is already fenced off these holds; close the manual path too.
-    if (prospect.quarantineReason === 'no_funded_external_buyer') {
-      throw new d.AppError(
-        'This lead is held for the MKTR Leads (external) buyer pool and cannot be manually assigned to an internal agent.',
-        409
-      );
-    }
-
-    // DNC fence: a lead held by the DNC gate must NOT be manually released (that would
-    // bypass scrubbing and hand a Do-Not-Call number to an adviser). It releases itself
-    // automatically once the DNC check clears (gate / backfill). assignProspect's claim
-    // below is reason-blind, so this guard is the fence.
-    if (prospect.quarantineReason === 'dnc_pending' || prospect.quarantineReason === 'dnc_registered') {
-      throw new d.AppError(
-        'This lead is held by the DNC (Do Not Call) gate and cannot be manually assigned — it releases automatically once the DNC check clears.',
-        409
-      );
-    }
-
-    // A manual admin assign is an EXEMPT route (decision a): it always delivers and does
-    // a best-effort deduct. If the prospect is currently HELD, this is a RELEASE — clear
-    // the hold ATOMICALLY (so a double-click / concurrent sweep can't deliver twice) and
-    // fire lead.assigned. Always lead.assigned, never lead.created: both receivers UPSERT
-    // on assigned (insert if unknown, re-point + un-hide if known), whereas a duplicate
-    // lead.created is a silent no-op — so a returned-to-held lead released via
-    // lead.created would never re-surface in the agent's app.
-    if (prospect.quarantinedAt) {
-      // Screening override bookkeeping (plan §9.5) — captured BEFORE the claim
-      // clears quarantineReason. A capture-charged, un-refunded screening lead
-      // must not be deducted AGAIN on release (Codex #2); a refunded one
-      // (screening_failed override) deducts normally like any release.
-      const screeningOverride = SCREENING_REASONS.includes(prospect.quarantineReason);
-      const screeningCaptureCharged =
-        screeningOverride &&
-        prospect.screeningMetadata?.alreadyCharged === true &&
-        prospect.screeningMetadata?.chargeRefunded !== true;
-
-      const [releaseRows] = await d.sequelize.query(
-        `UPDATE prospects
-            SET "assignedAgentId" = :agentId, "lastContactDate" = NOW(),
-                "quarantinedAt" = NULL, "quarantineReason" = NULL, "updatedAt" = NOW()
-          WHERE id = :prospectId AND "quarantinedAt" IS NOT NULL
-          RETURNING id`,
-        { replacements: { agentId, prospectId: prospect.id } }
-      );
-      const released = Array.isArray(releaseRows) && releaseRows.length > 0;
-      await prospect.reload();
-
-      if (!released) {
-        // Lost the race — already released elsewhere. Do not double-deliver, and return
-        // agent:null so the controller does not email an agent about a lead a concurrent
-        // release/sweep already assigned (possibly to a different agent).
-        return { prospect, agent: null, prospectWithCampaign: prospect };
-      }
-
-      await m.ProspectActivity.create({
-        prospectId: prospect.id,
-        type: 'assigned',
-        actorUserId: user?.id || null,
-        description: `Released from hold and assigned to ${agent.firstName} ${agent.lastName}`.trim(),
-        metadata: {
-          assignedAgentId: agentId,
-          previousAgentId,
-          released: true,
-          ...(screeningOverride ? { screeningOverride: true } : {}),
-        },
-      });
-
-      if (!screeningCaptureCharged) {
-        await d
-          .deductLeadCredit({ agentId, campaignId: prospect.campaignId || null })
-          .catch((err) => d.logger.error('Failed to deduct credit', { error: err?.message || String(err) }));
-      }
-
-      const prospectWithCampaign = await m.Prospect.findByPk(prospect.id, {
-        include: [
-          { association: 'campaign', attributes: ['id', 'name'] },
-          { association: 'qrTag', attributes: ['id', 'slug'] },
-        ],
-      });
-
-      const releaseDestination = destinationForAgent(agent);
-      d.dispatchEvent('lead.assigned', () =>
-        withBatchContext(
-          buildLeadAssignedPayload(prospect, agent, prospectWithCampaign, {
-            qrTag: prospectWithCampaign?.qrTag || null,
-            routingMode: 'direct',
-          }),
-          batch
-        ),
-        { destination: releaseDestination }
-      );
-
-      return { prospect, agent, prospectWithCampaign };
-    }
-
-    // ── Normal reassign (not held) ──
-    await prospect.update({
-      assignedAgentId: agentId,
-      lastContactDate: new Date(),
-    });
-
-    await m.ProspectActivity.create({
-      prospectId: prospect.id,
-      type: 'assigned',
-      actorUserId: user?.id || null,
-      description: `Assigned to agent ${agent.firstName} ${agent.lastName}`.trim(),
-      metadata: { assignedAgentId: agentId, previousAgentId },
-    });
-
-    await d
-      .deductLeadCredit({ agentId, campaignId: prospect.campaignId || null })
-      .catch((err) => d.logger.error('Failed to deduct credit', { error: err?.message || String(err) }));
-
-    const prospectWithCampaign = await m.Prospect.findByPk(prospect.id, {
-      include: [
-        { association: 'campaign', attributes: ['id', 'name'] },
-        { association: 'qrTag', attributes: ['id', 'slug'] },
-      ],
-    });
-
-    // Fire lead.assigned webhook to the NEW owner's app.
-    const newDestination = destinationForAgent(agent);
-    d.dispatchEvent(
-      'lead.assigned',
-      () =>
-        withBatchContext(
-          buildLeadAssignedPayload(prospect, agent, prospectWithCampaign, {
-            qrTag: prospectWithCampaign?.qrTag || null,
-            routingMode: 'direct',
-          }),
-          batch
-        ),
-      { destination: newDestination }
-    );
-
-    // Cross-app reassignment: if the PREVIOUS owner lived in a different app, that app
-    // still holds a copy of this lead that would otherwise linger. Tell it to release the
-    // lead (lead.unassigned -> the receiver marks it disputed). A SAME-app reassignment
-    // needs nothing extra — the receiver re-points the single shared row when it handles
-    // lead.assigned, so firing unassigned there would wrongly dispute the now-reassigned row.
-    if (previousAgentId && previousAgentId !== agentId) {
-      const prevAgent = await m.User.findByPk(previousAgentId, {
-        attributes: ['id', 'lyfeId', 'mktrLeadsId'],
-      });
-      const prevDestination = destinationForAgent(prevAgent);
-      if (prevDestination && prevDestination !== newDestination) {
-        const previousAgentExternalId = externalIdForDestination(prevAgent, prevDestination);
-        d.dispatchEvent('lead.unassigned', () => buildLeadUnassignedPayload(prospect, previousAgentExternalId), {
-          destination: prevDestination,
-        });
-      }
-    }
-
-    return { prospect, agent, prospectWithCampaign };
-  }
-
-  /**
-   * Bulk assign prospects to an agent. Returns { affectedCount, releasedCount, skipped, agent }
-   * — counts feed the controller's email side-effect and the UI's skip-accounting toast.
-   */
-  async function bulkAssignProspects(prospectIds, agentId, user) {
-    if (!prospectIds || !Array.isArray(prospectIds) || !agentId) {
-      throw new d.AppError('Prospect IDs array and agent ID are required', 400);
-    }
-    const requestedIds = [...new Set(prospectIds)];
-
-    const agent = await m.User.findOne({
-      where: {
-        id: agentId,
-        role: 'agent',
-        isActive: true,
-      },
-    });
-
-    if (!agent) {
-      throw new d.AppError('Invalid or inactive agent', 400);
-    }
-
-    // Pre-flight (bulk-only): assignment mutates rows first and delivers after, so refuse
-    // to run when delivery is IMPOSSIBLE (webhooks disabled / no subscriber tagged for the
-    // agent's app) — otherwise the whole batch would be stranded: assigned in MKTR, never
-    // surfaced in the agent's app. Transient send failures are fine (the persistent
-    // delivery queue retries); this guards misconfiguration only. A destination-less
-    // (local-only) agent passes: no delivery is expected, matching single-assign.
-    const newDestination = destinationForAgent(agent);
-    if (!(await d.hasDeliverableSubscriber('lead.assigned', newDestination))) {
-      throw new d.AppError(
-        "Lead delivery is not configured for this agent's app (webhooks disabled or no subscriber) — bulk assign would strand the leads. Fix webhook configuration and retry.",
-        409
-      );
-    }
-
-    const scopeFilter = await d.buildProspectWhere(user);
-    const whereConditions = {
-      id: { [Op.in]: requestedIds },
-      ...scopeFilter,
-      [Op.and]: [
-        // HELD rows are eligible only for the releasable reasons — bulk assign then acts
-        // as a RELEASE (the quarantine is cleared in the same atomic UPDATE below, and
-        // lead.assigned upserts at the receiver). Fenced reasons (external buyer pool,
-        // DNC gate) stay excluded, mirroring single-assign's guards.
-        { [Op.or]: [{ quarantinedAt: null }, { quarantineReason: { [Op.in]: RELEASABLE_HOLD_REASONS } }] },
-        // Skip rows already assigned to THIS agent (IS DISTINCT FROM semantics —
-        // a bare Op.ne would also exclude unassigned NULL rows). Re-assigning a
-        // no-op row used to double-charge the agent for a lead they already held.
-        { [Op.or]: [{ assignedAgentId: null }, { assignedAgentId: { [Op.ne]: agentId } }] },
-        // External-buyer-owned rows are fenced: assignedAgentId and
-        // externalAgentId are mutually exclusive owners, and writing the
-        // internal FK without clearing the external one would double-own the
-        // lead. Transfers from the external pool are not a bulk operation.
-        { externalAgentId: null },
-      ],
-    };
-
-    // Lock the candidate rows and update them inside ONE transaction so the webhook side
-    // sees a consistent picture: a concurrent (re)assignment of the same lead cannot slip
-    // between our read and our write (which would otherwise mis-attribute or skip a
-    // cross-app release and leave the other app holding an active copy). RETURNING stays the
-    // source of truth for WHICH rows changed, so per-campaign credit counting is exact. We
-    // lock WITHOUT the campaign include — FOR UPDATE cannot be applied to the nullable side
-    // of an outer join — and fetch campaign data for the payloads afterwards.
-    let result = [0, []];
-    const lockedById = new Map();
-    let requestedRows = [];
-    await d.sequelize.transaction(async (transaction) => {
-      const locked = await m.Prospect.findAll({
-        where: whereConditions,
-        // quarantineReason + screeningMetadata ride along so the deduct below
-        // can skip capture-charged screening releases (double-charge guard §9.5).
-        attributes: ['id', 'assignedAgentId', 'campaignId', 'quarantinedAt', 'quarantineReason', 'screeningMetadata'],
-        transaction,
-        lock: true,
-      });
-      for (const row of locked) lockedById.set(row.id, row);
-      result = await m.Prospect.update(
-        // Release + assign is ONE atomic write: clearing the hold here means a concurrent
-        // release (external dispatch / double-submit) blocks on the row lock, re-reads
-        // quarantinedAt IS NULL, and matches nothing — never a second delivery.
-        { assignedAgentId: agentId, lastContactDate: new Date(), quarantinedAt: null, quarantineReason: null },
-        // Update exactly the locked set; RETURNING reports the rows actually changed.
-        { where: { id: { [Op.in]: [...lockedById.keys()] } }, returning: ['id', 'campaignId'], transaction }
-      );
-      // Snapshot every requested (in-scope) row for skip classification, same txn so the
-      // classification matches what the locked UPDATE saw.
-      requestedRows = await m.Prospect.findAll({
-        where: { id: { [Op.in]: requestedIds }, ...scopeFilter },
-        attributes: ['id', 'assignedAgentId', 'externalAgentId', 'quarantinedAt', 'quarantineReason'],
-        transaction,
-      });
-    });
-
-    const affectedCount = result[0];
-    const affectedRows = result[1] || [];
-
-    // Skip accounting for the UI toast: classify every requested id that did NOT change.
-    // (Post-UPDATE snapshot: affected rows are classified off the pre-UPDATE lock set.)
-    const requestedById = new Map(requestedRows.map((r) => [r.id, r]));
-    const skipped = { notFound: 0, alreadyAssigned: 0, heldFenced: 0, externalOwned: 0 };
-    let releasedCount = 0;
-    for (const id of requestedIds) {
-      if (lockedById.has(id)) {
-        if (lockedById.get(id).quarantinedAt) releasedCount += 1;
-        continue;
-      }
-      const row = requestedById.get(id);
-      if (!row) skipped.notFound += 1;
-      else if (row.externalAgentId) skipped.externalOwned += 1;
-      else if (row.quarantinedAt && !RELEASABLE_HOLD_REASONS.includes(row.quarantineReason)) skipped.heldFenced += 1;
-      else skipped.alreadyAssigned += 1;
-    }
-    if (affectedCount > 0) {
-      const countsByCampaign = new Map();
-      for (const row of affectedRows) {
-        // Capture-charged, un-refunded screening leads already paid at capture
-        // — releasing them must not deduct a second credit (Codex #2).
-        const pre = lockedById.get(row.id);
-        const screeningCaptureCharged =
-          pre &&
-          SCREENING_REASONS.includes(pre.quarantineReason) &&
-          pre.screeningMetadata?.alreadyCharged === true &&
-          pre.screeningMetadata?.chargeRefunded !== true;
-        if (screeningCaptureCharged) continue;
-        const cId = row.campaignId || null;
-        countsByCampaign.set(cId, (countsByCampaign.get(cId) || 0) + 1);
-      }
-      for (const [cId, count] of countsByCampaign) {
-        await d
-          .deductLeadCredit({ agentId, campaignId: cId, amount: count })
-          .catch((err) => d.logger.error('Failed to deduct credits', { error: err?.message || String(err) }));
-      }
-
-      // Deliver each newly-assigned lead (bulk-assign previously fired NO webhook at all, so
-      // bulk-assigned leads never reached the agent's app). Mirror the single-assign path:
-      // lead.assigned to the new owner, plus — for a CROSS-app reassignment — lead.unassigned
-      // to the previous owner so its copy in the other app doesn't linger. Payload rows are
-      // fetched with their campaign; previous owners come from the locked snapshot.
-      // One batch context for the whole fan-out: the mktr-leads receiver coalesces the N
-      // per-lead pushes into a single "{size} leads assigned to you" summary (Lyfe ignores
-      // batch for now — per-lead pushes, the pre-batch behavior).
-      const batch = affectedCount > 1 ? { id: randomUUID(), size: affectedCount } : null;
-      const affectedIds = affectedRows.map((row) => row.id);
-      const full = await m.Prospect.findAll({
-        where: { id: { [Op.in]: affectedIds } },
-        include: [
-          { association: 'campaign', attributes: ['id', 'name'] },
-          { association: 'qrTag', attributes: ['id', 'slug'] },
-        ],
-      });
-
-      const prevOwnerIds = [
-        ...new Set(
-          affectedIds
-            .map((id) => lockedById.get(id)?.assignedAgentId)
-            .filter((prevId) => prevId && prevId !== agentId)
-        ),
-      ];
-      const prevAgentById = new Map();
-      if (prevOwnerIds.length > 0) {
-        const prevAgents = await m.User.findAll({
-          where: { id: { [Op.in]: prevOwnerIds } },
-          attributes: ['id', 'lyfeId', 'mktrLeadsId'],
-        });
-        for (const a of prevAgents) prevAgentById.set(a.id, a);
-      }
-
-      for (const p of full) {
-        d.dispatchEvent('lead.assigned', () =>
-          withBatchContext(
-            buildLeadAssignedPayload(p, agent, p, { qrTag: p.qrTag || null, routingMode: 'direct' }),
-            batch
-          ),
-          { destination: newDestination }
-        );
-
-        const prevId = lockedById.get(p.id)?.assignedAgentId;
-        const prevAgent = prevId && prevId !== agentId ? prevAgentById.get(prevId) : null;
-        if (prevAgent) {
-          const prevDestination = destinationForAgent(prevAgent);
-          if (prevDestination && prevDestination !== newDestination) {
-            const previousAgentExternalId = externalIdForDestination(prevAgent, prevDestination);
-            d.dispatchEvent('lead.unassigned', () => buildLeadUnassignedPayload(p, previousAgentExternalId), {
-              destination: prevDestination,
-            });
-          }
-        }
-      }
-
-      // Log a ProspectActivity per newly-assigned lead so BULK assignment lands on the unified
-      // timeline too — single-assign already logs (assignProspect), this path historically wrote
-      // none, so bulk-assigned leads were missing their "assigned" event. Best-effort (post-commit,
-      // like the credit deduction above) — never fail the assignment over an audit-row write.
-      await m.ProspectActivity.bulkCreate(
-        full.map((p) => {
-          const lockedRow = lockedById.get(p.id);
-          const prevId = lockedRow?.assignedAgentId;
-          return {
-            prospectId: p.id,
-            type: 'assigned',
-            actorUserId: user?.id || null,
-            description: `Assigned to ${agent.firstName} ${agent.lastName}`.trim(),
-            // Flag a reassignment (a prior owner existed) as a BOOLEAN so the timeline renders
-            // 'reassigned' — never expose who held it before. `released` marks a row that was
-            // HELD when the bulk assign claimed it (audit parity with single-release).
-            metadata: {
-              assignedAgentId: agent.id,
-              via: 'bulk_assign',
-              ...(prevId && prevId !== agentId ? { reassigned: true } : {}),
-              ...(lockedRow?.quarantinedAt ? { released: true } : {}),
-            },
-          };
-        })
-      ).catch((err) => d.logger.error('Failed to log bulk-assign activity', { error: err?.message || String(err) }));
-    }
-
-    return { affectedCount, releasedCount, skipped, agent };
-  }
-
-  /**
-   * Get prospect statistics for the user's scope.
-   */
-  async function getProspectStats(user) {
-    const whereConditions = await d.buildProspectWhere(user);
-
-    const [totalProspects, prospectsByStatus, prospectsBySource, prospectsByPriority, recentProspects, convertedCount] =
-      await Promise.all([
-        m.Prospect.count({ where: whereConditions }),
-        m.Prospect.findAll({
-          where: whereConditions,
-          attributes: ['leadStatus', [d.sequelize.fn('COUNT', d.sequelize.col('leadStatus')), 'count']],
-          group: ['leadStatus'],
-        }),
-        m.Prospect.findAll({
-          where: whereConditions,
-          attributes: ['leadSource', [d.sequelize.fn('COUNT', d.sequelize.col('leadSource')), 'count']],
-          group: ['leadSource'],
-        }),
-        m.Prospect.findAll({
-          where: whereConditions,
-          attributes: ['priority', [d.sequelize.fn('COUNT', d.sequelize.col('priority')), 'count']],
-          group: ['priority'],
-        }),
-        m.Prospect.findAll({
-          where: {
-            ...whereConditions,
-            createdAt: {
-              [Op.gte]: new Date(Date.now() - 7 * 24 * 60 * 60 * 1000),
-            },
-          },
-          limit: 10,
-          order: [['createdAt', 'DESC']],
-          attributes: ['id', 'firstName', 'lastName', 'email', 'leadStatus', 'createdAt'],
-          include: [
-            {
-              association: 'campaign',
-              attributes: ['id', 'name'],
-            },
-          ],
-        }),
-        m.Prospect.count({
-          where: { ...whereConditions, leadStatus: 'won' },
-        }),
-      ]);
-    const conversionRate = totalProspects > 0 ? ((convertedCount / totalProspects) * 100).toFixed(2) : 0;
-
-    return {
-      totalProspects,
-      conversionRate: parseFloat(conversionRate),
-      byStatus: prospectsByStatus.map((item) => ({
-        status: item.leadStatus,
-        count: parseInt(item.dataValues.count),
-      })),
-      bySource: prospectsBySource.map((item) => ({
-        source: item.leadSource,
-        count: parseInt(item.dataValues.count),
-      })),
-      byPriority: prospectsByPriority.map((item) => ({
-        priority: item.priority,
-        count: parseInt(item.dataValues.count),
-      })),
-      recentProspects,
-    };
-  }
-
-  /**
-   * List prospects with pagination, filtering, and auth scoping.
-   */
-  async function listProspects(user, params) {
-    const {
-      page = 1,
-      limit = 10,
-      leadStatus,
-      priority,
-      leadSource,
-      assignedAgentId,
-      assignment,
-      campaignId,
-      search,
-      dateFrom,
-      dateTo,
-      qrTagId,
-      sort,
-    } = params;
-
-    // Clamp pagination so malformed query params (e.g. ?page=-1&limit=-5) don't
-    // reach Sequelize as a negative LIMIT/OFFSET, which throws → 500.
-    const pageNum = Math.max(1, parseInt(page, 10) || 1);
-    const limitNum = Math.min(Math.max(1, parseInt(limit, 10) || 10), 200);
-    const offset = (pageNum - 1) * limitNum;
-    const scopeFilter = await d.buildProspectWhere(user);
-    const whereConditions = { ...scopeFilter };
-
-    // leadStatus / leadSource accept comma-lists (Phase B multi-select filters);
-    // single values stay backward-compatible. Tokens are whitelisted against
-    // the enum so nothing unknown ever reaches a Postgres enum cast (→ 500);
-    // an all-invalid value degrades to an empty page — the long-standing
-    // single-value behavior.
-    const statusFilter = parseEnumFilter(leadStatus, VALID_LEAD_STATUSES);
-    const sourceFilter = parseEnumFilter(leadSource, VALID_LEAD_SOURCES);
-    if (statusFilter.empty || sourceFilter.empty) {
-      return { prospects: [], pagination: { currentPage: pageNum, totalPages: 0, totalItems: 0, itemsPerPage: limitNum } };
-    }
-    if (assignment && !VALID_ASSIGNMENT_FILTERS.includes(assignment)) {
-      return { prospects: [], pagination: { currentPage: pageNum, totalPages: 0, totalItems: 0, itemsPerPage: limitNum } };
-    }
-
-    // Assignment-state filter: 'unassigned' means truly in limbo (not held — held rows
-    // have their own bucket so the admin's pending pool and the strays stay separable).
-    // Prospects carry TWO mutually-exclusive assignee FKs (assignedAgentId for users,
-    // externalAgentId for the ExternalAgent buyer pool) — both count as "assigned".
-    // The Op.or lives inside Op.and so it can never clobber the search Op.or below.
-    if (assignment === 'assigned') {
-      whereConditions[Op.and] = [
-        ...(whereConditions[Op.and] || []),
-        { [Op.or]: [{ assignedAgentId: { [Op.ne]: null } }, { externalAgentId: { [Op.ne]: null } }] },
-      ];
-    } else if (assignment === 'unassigned') {
-      whereConditions.assignedAgentId = null;
-      whereConditions.externalAgentId = null;
-      whereConditions.quarantinedAt = null;
-    } else if (assignment === 'held') whereConditions.quarantinedAt = { [Op.ne]: null };
-
-    if (qrTagId) whereConditions.qrTagId = qrTagId;
-    if (statusFilter.values) {
-      whereConditions.leadStatus = statusFilter.values.length === 1 ? statusFilter.values[0] : { [Op.in]: statusFilter.values };
-    }
-    if (priority) whereConditions.priority = priority;
-    if (sourceFilter.values) {
-      whereConditions.leadSource = sourceFilter.values.length === 1 ? sourceFilter.values[0] : { [Op.in]: sourceFilter.values };
-    }
-    if (assignedAgentId) whereConditions.assignedAgentId = assignedAgentId;
-    if (campaignId) whereConditions.campaignId = campaignId;
-
-    if (search) {
-      const sanitizedSearch = String(search).slice(0, 100).replace(/%/g, '\\%').replace(/_/g, '\\_');
-      const likeOp = Op.iLike;
-      whereConditions[Op.or] = [
-        { firstName: { [likeOp]: `%${sanitizedSearch}%` } },
-        { lastName: { [likeOp]: `%${sanitizedSearch}%` } },
-        { email: { [likeOp]: `%${sanitizedSearch}%` } },
-        { company: { [likeOp]: `%${sanitizedSearch}%` } },
-        // Phone matches ignore spacing (stored "+65 9231 8804", operators type
-        // digits) — strip spaces on both sides of the comparison. Sequelize
-        // STATICS (not instance methods) so DI-mocked unit suites stay valid.
-        sqlWhere(
-          sqlFn('REPLACE', sqlCol('Prospect.phone'), ' ', ''),
-          { [likeOp]: `%${sanitizedSearch.replace(/\s+/g, '')}%` }
-        ),
-      ];
-    }
-
-    if (dateFrom || dateTo) {
-      whereConditions.createdAt = {};
-      if (dateFrom) whereConditions.createdAt[Op.gte] = new Date(dateFrom);
-      if (dateTo) whereConditions.createdAt[Op.lte] = new Date(dateTo);
-    }
-
-    const { count, rows: prospects } = await m.Prospect.findAndCountAll({
-      where: whereConditions,
-      // Screening evidence (call transcripts/summaries) is detail-only — never
-      // shipped in list projections (plan §4, Codex #16).
-      attributes: { exclude: ['screeningMetadata'] },
-      limit: limitNum,
-      offset,
-      order: parseProspectSort(sort),
-      include: [
-        {
-          association: 'assignedAgent',
-          attributes: ['id', 'firstName', 'lastName', 'email'],
-        },
-        {
-          association: 'campaign',
-          attributes: ['id', 'name', 'type', 'status'],
-        },
-        {
-          association: 'qrTag',
-          attributes: ['id', 'name', 'type'],
-        },
-      ],
-    });
-
-    // Admin-only: attach a light cross-campaign repeat-signup count per row for
-    // the list badge. Non-admins never receive it (this list endpoint is shared
-    // with the agent MyProspects view). Resilient — failure → no badge data.
-    if (user?.role === 'admin' && prospects.length > 0) {
-      const counts = await repeatSignupCounts(
-        d.sequelize,
-        prospects.map((p) => ({ id: p.id, phone: p.phone, email: p.email }))
-      ).catch(() => new Map());
-      for (const p of prospects) p.setDataValue('repeatSignupCount', counts.get(p.id) ?? null);
-    }
-
-    // Admin-only, OPT-IN (?include=outcome — the v2 list's STATUS column):
-    // per-row campaign outcome, the same contract discipline as
-    // ?include=profile. Byte-identical without the flag; agent MyProspects
-    // and the palette never pay for it. Resilient — a miss drops the chips,
-    // never the page.
-    const wantOutcome = user?.role === 'admin'
-      && String(params.include || '').split(',').map((s2) => s2.trim()).includes('outcome')
-      && prospects.length > 0;
-    if (wantOutcome) {
-      const outcomes = await d.getProspectOutcomes(prospects).catch(() => new Map());
-      for (const p of prospects) {
-        const o = outcomes.get(String(p.id));
-        if (o) {
-          p.setDataValue('draw', o.draw);
-          p.setDataValue('reward', o.reward);
-        }
-      }
-    }
-
-    return {
-      prospects,
-      pagination: {
-        currentPage: pageNum,
-        totalPages: Math.ceil(count / limitNum),
-        totalItems: count,
-        itemsPerPage: limitNum,
-      },
-    };
-  }
-
-  /**
    * Schedule a follow-up for a prospect.
    */
   async function scheduleFollowUp(id, { nextFollowUpDate, notes }, user) {
@@ -2435,547 +1657,6 @@ export function makeProspectService(overrides = {}) {
   }
 
   /**
-   * Track a prospect view.
-   */
-  async function trackProspectView(id, user, { source, userAgent } = {}) {
-    const scopeWhere = await d.buildProspectWhere(user);
-    const prospect = await m.Prospect.findOne({ where: { id, ...scopeWhere } });
-
-    if (!prospect) {
-      throw new d.AppError('Prospect not found or access denied', 404);
-    }
-
-    await m.ProspectActivity.create({
-      prospectId: prospect.id,
-      type: 'viewed',
-      actorUserId: user.id,
-      description: `Prospect viewed by ${user.firstName || 'agent'} ${user.lastName || ''}`,
-      metadata: {
-        source: source || 'email_link',
-        viewedAt: new Date(),
-        userAgent,
-      },
-    });
-  }
-
-  /**
-   * List HELD (quarantined) leads, FIFO by quarantinedAt (the release order). Scoped to
-   * the caller's access; optionally filtered by campaign. Release is done via the
-   * existing assignProspect (PATCH /:id/assign) or the auto-release sweep on top-up.
-   */
-  async function listHeldProspects(user, params = {}) {
-    const { campaignId, quarantineReason } = params;
-    const limit = Math.min(parseInt(params.limit, 10) || 100, 500);
-    const scopeFilter = await d.buildProspectWhere(user);
-    const where = { ...scopeFilter, quarantinedAt: { [Op.ne]: null } };
-    if (campaignId) where.campaignId = campaignId;
-    // Filter by reason IN the query (before the limit) so assignable holds are never
-    // hidden behind a page of external-buyer holds.
-    if (quarantineReason) where.quarantineReason = quarantineReason;
-
-    const { count, rows } = await m.Prospect.findAndCountAll({
-      where,
-      attributes: { exclude: ['screeningMetadata'] },
-      include: [{ association: 'campaign', attributes: ['id', 'name'] }],
-      order: [['quarantinedAt', 'ASC']], // FIFO — oldest held first (the release order)
-      limit,
-    });
-
-    const held = rows.map((p) => ({
-      id: p.id,
-      firstName: p.firstName,
-      lastName: p.lastName,
-      phone: p.phone,
-      email: p.email,
-      leadSource: p.leadSource,
-      campaignId: p.campaignId,
-      campaignName: p.campaign?.name || null,
-      quarantinedAt: p.quarantinedAt,
-      quarantineReason: p.quarantineReason,
-      createdAt: p.createdAt,
-    }));
-
-    return { count, held };
-  }
-
-  // The System-Agent routing fallback is only a true orphan marker when it has NO delivery
-  // destination (no lyfeId / mktrLeadsId) — its leads were never delivered anywhere. If
-  // DEFAULT_AGENT_ID ever points at a REAL fallback agent (whose leads DO get delivered),
-  // this returns null so those leads can never be mistaken for orphans (Codex B/P1).
-  async function orphanSystemAgentId() {
-    const id = await d.getSystemAgentId();
-    if (!id) return null;
-    const u = await m.User.findByPk(id, { attributes: ['id', 'lyfeId', 'mktrLeadsId'] });
-    return u && !u.lyfeId && !u.mktrLeadsId ? id : null;
-  }
-
-  /**
-   * Fleet-wide list of dispatchable ORPHANS for the external admin queue: no_funded_agent
-   * HOLDS *and* leads parked on the phantom System Agent (the soft-campaign fallback,
-   * which has no phone so they were never delivered). Each row is tagged with `reason`
-   * ('no_funded_agent' | 'unassigned') and a `since` timestamp.
-   */
-  async function listDispatchableOrphans({ campaignId = null, limit } = {}) {
-    const systemAgentId = await orphanSystemAgentId();
-    const lim = Math.min(parseInt(limit, 10) || 50, 200);
-    const orphanClauses = [{ quarantinedAt: { [Op.ne]: null }, quarantineReason: 'no_funded_agent' }];
-    if (systemAgentId) orphanClauses.push({ assignedAgentId: systemAgentId, quarantinedAt: null });
-    const where = { [Op.or]: orphanClauses };
-    if (campaignId) where.campaignId = campaignId;
-
-    const { count, rows } = await m.Prospect.findAndCountAll({
-      where,
-      attributes: { exclude: ['screeningMetadata'] },
-      include: [
-        { association: 'campaign', attributes: ['id', 'name'] },
-        // qrTag drives the "QR code" source label for a bound-QR lead whose leadSource isn't
-        // 'qr_code'. Select only REAL columns — QrTag has `slug` but NO `externalId` (that field
-        // exists on the webhook payload's qrTag, not the model), so signupSourceLabel keys off slug.
-        { association: 'qrTag', attributes: ['id', 'slug'] },
-      ],
-      order: [['createdAt', 'ASC']], // FIFO by signup (held + unassigned interleaved)
-      limit: lim,
-    });
-
-    const orphans = rows.map((p) => {
-      // Full personal / firmographic data the admin detail view shows (DOB, postal,
-      // company, …) — parity with a delivered lead, plus the demographics it drops.
-      const { birthday, details } = buildHeldLeadEnrichment(p);
-      return {
-        id: p.id,
-        firstName: p.firstName,
-        lastName: p.lastName,
-        phone: p.phone,
-        email: p.email,
-        leadSource: p.leadSource,
-        // Rich, delivered-lead-equivalent label ("Instagram ad", "Meta ad", "Web form", …)
-        // derived from sourceMetadata.utm — so the held queue reads the same source the lead
-        // will show once assigned, not the coarse capture channel.
-        sourceLabel: signupSourceLabel({ leadSource: p.leadSource, qrTag: p.qrTag, sourceMetadata: p.sourceMetadata }),
-        // Raw DOB string — the buyer app formats it to DD/MM/YYYY (one formatter, app-side).
-        birthday,
-        // Ordered, display-ready [{ label, value }] enrichment rows (no PII beyond what the
-        // admin-only detail view already shows; never enters the non-PII summary projection).
-        details,
-        campaignId: p.campaignId,
-        campaignName: p.campaign?.name || null,
-        reason: p.quarantinedAt ? 'no_funded_agent' : 'unassigned',
-        since: p.quarantinedAt || p.createdAt,
-        createdAt: p.createdAt,
-      };
-    });
-
-    return { count, orphans };
-  }
-
-  /**
-   * Assign an ORPHANED prospect (a no_funded_agent HOLD, or a lead parked on the phantom
-   * System Agent fallback) to a mktr-leads agent and deliver it via a fresh `lead.assigned`
-   * — the lead is new to the destination app, so the receiver INSERTs it from that payload
-   * AND records an explicit "assigned @ now" activity in the lead's timeline.
-   *
-   * The held-only counterpart to assignProspect, purpose-built for the external
-   * mktr-leads admin dispatch endpoint. Unlike assignProspect it can NEVER fall
-   * into the normal-reassignment path: a request that arrives after the lead is
-   * already released returns `already_handled` (no second charge, no second
-   * delivery). The release UPDATE and the `lead.assigned` delivery row are
-   * written in ONE transaction (outbox), so a crash can never leave a lead
-   * un-held yet undelivered (recoverPendingRetries flushes the row on restart).
-   *
-   * Agent identity is resolved by `mktrLeadsId` ONLY — never by phone, which can
-   * resolve a Lyfe-owned or provenance-less user whose webhook destination would
-   * not be mktr_leads. A manual admin assign is an explicit override: it always
-   * delivers and only best-effort deducts a campaign credit (the lead was held
-   * precisely because nobody was funded).
-   *
-   * @returns {Promise<{status:'assigned'|'already_handled'|'invalid_agent'|'not_found'|'not_assignable_external', leadId?:string, agent?:object}>}
-   */
-  async function releaseHeldProspect(prospectId, agentMktrLeadsId, opts = {}) {
-    const { idempotencyKey = null, actorUserId = null, batch = null } = opts;
-    const IDEMP_SCOPE = 'external:held-assign';
-
-    // Replay: a retried request with the same key returns the first result verbatim.
-    {
-      const replay = await idem.replayIfDone(IDEMP_SCOPE, idempotencyKey);
-      if (replay) return replay;
-    }
-
-    // Load the prospect FIRST so a retry after a successful release reports
-    // already_handled — not invalid_agent if the chosen agent was since deactivated.
-    const prospect = await m.Prospect.findByPk(prospectId);
-    if (!prospect) return { status: 'not_found' };
-    // External-buyer holds can never be manually released to an internal agent.
-    if (prospect.quarantineReason === 'no_funded_external_buyer') {
-      return { status: 'not_assignable_external' };
-    }
-    // An orphan = a lead with no real owner: a no_funded_agent HOLD, or one parked on
-    // the phantom System Agent (the soft-campaign fallback — never delivered anywhere).
-    // Anything else is a real assigned lead and must not be touched here.
-    const systemAgentId = await orphanSystemAgentId();
-    const isHeld = !!prospect.quarantinedAt && prospect.quarantineReason === 'no_funded_agent';
-    const isUnassigned = !prospect.quarantinedAt && !!systemAgentId && prospect.assignedAgentId === systemAgentId;
-    if (!isHeld && !isUnassigned) return { status: 'already_handled' };
-
-    // Resolve the destination agent by mktrLeadsId ONLY (phone is unsafe — it can
-    // resolve a Lyfe-owned / provenance-less user whose destination ≠ mktr_leads).
-    const agent = agentMktrLeadsId
-      ? await m.User.findOne({ where: { mktrLeadsId: agentMktrLeadsId, role: 'agent', isActive: true } })
-      : null;
-    if (!agent) return { status: 'invalid_agent' };
-
-    // Atomic release + delivery intent (transactional outbox).
-    const t = await d.sequelize.transaction();
-    let result;
-    let deliveryPairs = [];
-    try {
-      // Held-only conditional release: gated on the row STILL being a no_funded_agent
-      // hold, so a racing duplicate or the auto sweep loses the row lock and sees
-      // `quarantinedAt IS NULL` → 0 rows → already_handled (never a second delivery).
-      const [rows] = await d.sequelize.query(
-        `UPDATE prospects
-            SET "assignedAgentId" = :agentId, "lastContactDate" = NOW(),
-                "quarantinedAt" = NULL, "quarantineReason" = NULL, "updatedAt" = NOW()
-          WHERE id = :prospectId AND (
-            ("quarantinedAt" IS NOT NULL AND "quarantineReason" = 'no_funded_agent')
-            OR ("assignedAgentId" = :systemAgentId AND "quarantinedAt" IS NULL)
-          )
-          RETURNING id`,
-        { replacements: { agentId: agent.id, prospectId, systemAgentId }, transaction: t }
-      );
-
-      if (!Array.isArray(rows) || rows.length === 0) {
-        result = { status: 'already_handled' };
-      } else {
-        await m.ProspectActivity.create({
-          prospectId,
-          type: 'assigned',
-          actorUserId,
-          description: `Assigned to ${agent.firstName} ${agent.lastName} via the dispatch queue`.trim(),
-          metadata: { assignedAgentId: agent.id, released: true, via: 'external_app', fromSystemAgent: isUnassigned },
-        }, { transaction: t });
-
-        const withCampaign = await m.Prospect.findByPk(prospectId, {
-          include: [
-            { association: 'campaign', attributes: ['id', 'name'] },
-            { association: 'qrTag', attributes: ['id', 'slug'] },
-          ],
-          transaction: t,
-        });
-
-        // Destination is mktr_leads (agent has mktrLeadsId set); the receiver matches
-        // `routing.agentExternalId` against agents.mktr_user_id. buildLeadAssignedPayload sets
-        // that id via externalIdForDestination(agent,'mktr_leads') === agent.mktrLeadsId.
-        const destination = destinationForAgent(agent);
-
-        // Fire lead.assigned (NOT lead.created) so the mktr-leads receiver records an explicit
-        // "assigned @ now" activity on top of "received @ signup" — the dispatched lead's timeline
-        // then shows the assignment AND its time, matching the MKTR-side activity logged above.
-        // The lead is brand-new to the destination app, so the receiver INSERTs it from this
-        // payload (the proven lead.assigned-inserts-new path); the lead block is identical to
-        // buildLeadCreatedPayload, so delivery is otherwise unchanged.
-        deliveryPairs = await d.persistEventDeliveries(
-          'lead.assigned',
-          () =>
-            withBatchContext(
-              buildLeadAssignedPayload(withCampaign, agent, withCampaign, {
-                qrTag: withCampaign?.qrTag || null,
-                routingMode: 'direct',
-              }),
-              batch
-            ),
-          { destination },
-          t
-        );
-        // Fail closed: NEVER release a lead we cannot durably deliver. An empty set
-        // means webhooks are disabled or no subscriber is tagged for this
-        // destination — roll back so the lead stays held instead of vanishing.
-        if (deliveryPairs.length === 0) {
-          await t.rollback();
-          return { status: 'undeliverable' };
-        }
-
-        result = { status: 'assigned', leadId: prospectId, agent: { firstName: agent.firstName, lastName: agent.lastName } };
-      }
-
-      // Record idempotency atomically with the release so an exact retry replays this
-      // result verbatim. A concurrent same-key duplicate loses the unique PK and rolls
-      // back — harmless, since the held-only release already prevents any double effect.
-      await idem.recordResult(IDEMP_SCOPE, idempotencyKey, result, { transaction: t });
-
-      await t.commit();
-    } catch (err) {
-      await t.rollback();
-      // A concurrent request with the SAME idempotency key won the unique PK — replay
-      // its recorded result instead of surfacing a 500 (the held-only release already
-      // guaranteed no double effect).
-      {
-        const winner = await idem.replayOnUniqueRace(IDEMP_SCOPE, idempotencyKey, err);
-        if (winner) return winner;
-      }
-      throw err;
-    }
-
-    // Post-commit side-effects — never block or roll back the durable release.
-    if (result.status === 'assigned') {
-      await d.deductLeadCredit({ agentId: agent.id, campaignId: prospect.campaignId || null })
-        .catch((err) => d.logger.error('[releaseHeldProspect] credit deduct failed', { error: err?.message || String(err) }));
-      d.flushDeliveries(deliveryPairs);
-    }
-
-    return result;
-  }
-
-  /**
-   * Reassign an ASSIGNED lead to a different mktr-leads agent (admin, from the app). Wraps
-   * assignProspect (which fires lead.assigned → the mktr-leads receiver re-points the single
-   * shared row, so the PREVIOUS agent loses RLS access — same-app reassign fires no disputed
-   * lead.unassigned). Idempotent: a retry with the same key replays; re-targeting the SAME agent
-   * is a no-op (no double charge). NOT for held/orphan leads — those use releaseHeldProspect.
-   *
-   * @returns {Promise<{status:'reassigned'|'invalid_agent'|'not_found'|'not_assignable', leadId?:string, agent?:object}>}
-   */
-  async function reassignProspectExternal(prospectId, agentMktrLeadsId, opts = {}) {
-    const { idempotencyKey = null, actorUserId = null, batch = null } = opts;
-    const IDEMP_SCOPE = 'external:admin-reassign';
-
-    // Claim the idempotency key FIRST (unique PK) so concurrent / retried same-key requests can
-    // never both run assignProspect (which always charges + dispatches). A completed claim replays
-    // its result; a still-running claim reports in_progress (the caller retries).
-    {
-      const claim = await idem.claimOrReplay(IDEMP_SCOPE, idempotencyKey);
-      if (!claim.claimed) return claim.replay;
-    }
-
-    // Record the outcome (success OR validation failure) on the claimed key so a retry replays it.
-    const record = (result) => idem.recordClaimed(IDEMP_SCOPE, idempotencyKey, result);
-
-    // Resolve the destination by mktrLeadsId ONLY (excludes Lyfe / provenance-less users).
-    const agent = agentMktrLeadsId
-      ? await m.User.findOne({ where: { mktrLeadsId: agentMktrLeadsId, role: 'agent', isActive: true } })
-      : null;
-    if (!agent) return record({ status: 'invalid_agent' });
-
-    const prospect = await m.Prospect.findByPk(prospectId);
-    if (!prospect) return record({ status: 'not_found' });
-    // Only a lead currently assigned to a real agent is reassignable here; held/orphan leads go
-    // through the held queue (releaseHeldProspect) so we never double-handle them.
-    if (prospect.quarantinedAt || !prospect.assignedAgentId) return record({ status: 'not_assignable' });
-
-    // Same-app scope: the lead's CURRENT owner must be a mktr-leads agent (a Lyfe-owned lead would
-    // route through the Lyfe app — out of scope, and its receiver isn't wired for this).
-    const currentOwner = await m.User.findByPk(prospect.assignedAgentId, { attributes: ['id', 'lyfeId', 'mktrLeadsId'] });
-    if (destinationForAgent(currentOwner) !== 'mktr_leads') return record({ status: 'not_assignable' });
-
-    const agentBrief = { firstName: agent.firstName, lastName: agent.lastName };
-    try {
-      // Already with the target → no-op so a retry / re-pick never double-charges the agent.
-      if (prospect.assignedAgentId !== agent.id) {
-        await assignProspect(prospectId, agent.id, { id: actorUserId }, { batch });
-      }
-    } catch (err) {
-      // assignProspect isn't transactional; record the failure so a retry replays it (no
-      // re-charge) and RETURN it as a typed result — a throw surfaces as a generic 500 and
-      // the caller's app would misbucket the FIRST response as a retryable transport
-      // failure instead of the sticky needs-attention state the recorded replay enforces.
-      d.logger.error('[external-reassign] assignProspect failed', {
-        prospectId,
-        error: err?.message || String(err),
-      });
-      return record({ status: 'error', error: 'reassign_failed' });
-    }
-    return record({ status: 'reassigned', leadId: prospectId, agent: agentBrief });
-  }
-
-  /**
-   * Return an ASSIGNED lead to the held queue (admin pull-back). Re-holds the prospect and
-   * fires lead.unassigned WITH `returnedToHeld` so the previous agent's app drops it — the
-   * mktr-leads receiver SOFT-DELETES (vanishes) the lead instead of disputing it; the Lyfe
-   * receiver nulls its assigned_to (the agent loses RLS visibility — same effect, flag
-   * ignored). Admins keep it, and a later dispatch (lead.assigned upsert) re-surfaces it.
-   * NO refund (consistent with reassign). Idempotent; fail-closed for deliverable owners
-   * (never re-hold a lead we can't durably vanish).
-   *
-   * Two callers, two flavors:
-   * - EXTERNAL (mktr-leads admin app, defaults): reason 'no_funded_agent' so the existing
-   *   external held queue + releaseHeldProspect handle it; mktr-leads-owned leads only.
-   * - WEB ADMIN (bulk return, opts): reason 'returned_by_admin' — deliberately invisible
-   *   to the external queue/release/sweep (all filter 'no_funded_agent'), so a returned
-   *   Lyfe-owned lead can never leak to the external buyer pool. `anyDestination` admits
-   *   Lyfe-owned and no-destination owners (System Agent — nothing was ever delivered, so
-   *   the vanish is skipped and fail-closed does not apply). `promoteUnassigned` folds
-   *   already-unassigned strays into the held pool (no webhook — nothing to vanish).
-   *
-   * @returns {Promise<{status:'returned'|'promoted'|'already_handled'|'not_assignable'|'not_found'|'undeliverable', leadId?:string}>}
-   */
-  async function returnProspectToHeld(prospectId, opts = {}) {
-    const {
-      idempotencyKey = null,
-      actorUserId = null,
-      reason = 'no_funded_agent',
-      via = 'external_app',
-      anyDestination = false,
-      promoteUnassigned = false,
-      scopeWhere = null,
-    } = opts;
-    const IDEMP_SCOPE = 'external:admin-return-held';
-
-    {
-      const replay = await idem.replayIfDone(IDEMP_SCOPE, idempotencyKey);
-      if (replay) return replay;
-    }
-
-    const prospect = await m.Prospect.findOne({ where: { id: prospectId, ...(scopeWhere || {}) } });
-    if (!prospect) return { status: 'not_found' };
-    const previousAgentId = prospect.assignedAgentId;
-    // A held lead is already in a queue → already_handled (a retry after a successful
-    // return replays as a no-op).
-    if (prospect.quarantinedAt) return { status: 'already_handled' };
-
-    // ── Promotion arm (web-admin only): an unassigned, unheld stray joins the held pool.
-    // No webhook — it has no owner app copy to vanish. Conditional UPDATE = race-safe.
-    if (!previousAgentId) {
-      // External-buyer-owned rows are NOT strays: externalAgentId is the other
-      // mutually-exclusive owner FK, and quarantining such a row would hold a
-      // lead an external buyer already paid for and holds a copy of.
-      if (prospect.externalAgentId) return { status: 'undeliverable' };
-      if (!promoteUnassigned) return { status: 'already_handled' };
-      const pt = await d.sequelize.transaction();
-      try {
-        const [rows] = await d.sequelize.query(
-          `UPDATE prospects
-              SET "quarantinedAt" = NOW(), "quarantineReason" = :reason, "updatedAt" = NOW()
-            WHERE id = :prospectId AND "assignedAgentId" IS NULL AND "quarantinedAt" IS NULL
-            RETURNING id`,
-          { replacements: { prospectId, reason }, transaction: pt }
-        );
-        if (!Array.isArray(rows) || rows.length === 0) {
-          await pt.rollback();
-          return { status: 'already_handled' };
-        }
-        await m.ProspectActivity.create({
-          prospectId,
-          type: 'assigned',
-          actorUserId,
-          description: 'Moved to held queue by admin',
-          metadata: { promoted: true, via },
-        }, { transaction: pt });
-        await pt.commit();
-        return { status: 'promoted', leadId: prospectId };
-      } catch (err) {
-        await pt.rollback();
-        throw err;
-      }
-    }
-
-    const prevAgent = await m.User.findByPk(previousAgentId, { attributes: ['id', 'lyfeId', 'mktrLeadsId'] });
-    const prevDestination = destinationForAgent(prevAgent);
-    // External flavor stays same-app-scoped: only a mktr-leads-owned lead can be returned
-    // (its broker + receiver own that contract). The web-admin flavor admits any owner.
-    if (!anyDestination && prevDestination !== 'mktr_leads') return { status: 'not_assignable' };
-    const previousAgentExternalId = externalIdForDestination(prevAgent, prevDestination);
-
-    const t = await d.sequelize.transaction();
-    let result;
-    let deliveryPairs = [];
-    try {
-      // Conditional re-hold: gated on the row STILL being assigned to this agent and not held, so
-      // a racing duplicate loses the row lock → 0 rows → already_handled (never a second vanish).
-      const [rows] = await d.sequelize.query(
-        `UPDATE prospects
-            SET "assignedAgentId" = NULL, "quarantinedAt" = NOW(),
-                "quarantineReason" = :reason, "updatedAt" = NOW()
-          WHERE id = :prospectId AND "assignedAgentId" = :previousAgentId AND "quarantinedAt" IS NULL
-          RETURNING id`,
-        { replacements: { prospectId, previousAgentId, reason }, transaction: t }
-      );
-
-      if (!Array.isArray(rows) || rows.length === 0) {
-        result = { status: 'already_handled' };
-      } else {
-        await m.ProspectActivity.create({
-          prospectId,
-          type: 'assigned',
-          actorUserId,
-          description: 'Returned to held queue by admin',
-          metadata: { previousAgentId, returnedToHeld: true, via },
-        }, { transaction: t });
-
-        if (prevDestination) {
-          // Vanish the lead from the previous agent's app — the mktr-leads receiver
-          // soft-deletes on returnedToHeld; Lyfe nulls assigned_to.
-          deliveryPairs = await d.persistEventDeliveries(
-            'lead.unassigned',
-            () => buildLeadUnassignedPayload(prospect, previousAgentExternalId, { returnedToHeld: true }),
-            { destination: prevDestination },
-            t
-          );
-          // Fail closed: never re-hold a lead we cannot durably vanish from the agent's app.
-          if (deliveryPairs.length === 0) {
-            await t.rollback();
-            return { status: 'undeliverable' };
-          }
-        }
-        // No-destination owner (System Agent / provenance-less): nothing was ever
-        // delivered, so there is nothing to vanish — the re-hold alone is complete.
-
-        result = { status: 'returned', leadId: prospectId };
-      }
-
-      await idem.recordResult(IDEMP_SCOPE, idempotencyKey, result, { transaction: t });
-
-      await t.commit();
-    } catch (err) {
-      await t.rollback();
-      {
-        const winner = await idem.replayOnUniqueRace(IDEMP_SCOPE, idempotencyKey, err);
-        if (winner) return winner;
-      }
-      throw err;
-    }
-
-    // Post-commit: flush the vanish delivery. NO credit refund (consistent with reassign).
-    if (result.status === 'returned') {
-      d.flushDeliveries(deliveryPairs);
-    }
-
-    return result;
-  }
-
-  /**
-   * Bulk return leads to the held queue (web admin). Fan-out over the hardened single op
-   * — the same pattern the mktr-leads admin app uses for its bulk — because per-row
-   * fail-closed vanish semantics don't fit one set-based transaction: each row either
-   * fully returns (re-hold + vanish delivery committed together) or reports why not.
-   * `returned_by_admin` keeps every returned lead OUT of the external buyer queue.
-   * No idempotency keys: the single op's conditional UPDATE makes re-runs no-ops.
-   */
-  async function bulkReturnProspectsToHeld(prospectIds, user) {
-    if (!prospectIds || !Array.isArray(prospectIds) || prospectIds.length === 0) {
-      throw new d.AppError('Prospect IDs array is required', 400);
-    }
-    const requestedIds = [...new Set(prospectIds)];
-    const scopeWhere = await d.buildProspectWhere(user);
-
-    const counts = { returned: 0, promoted: 0, alreadyHeld: 0, undeliverable: 0, notFound: 0 };
-    for (const id of requestedIds) {
-      const r = await returnProspectToHeld(id, {
-        actorUserId: user?.id || null,
-        reason: 'returned_by_admin',
-        via: 'web_admin',
-        anyDestination: true,
-        promoteUnassigned: true,
-        scopeWhere,
-      });
-      if (r.status === 'returned') counts.returned += 1;
-      else if (r.status === 'promoted') counts.promoted += 1;
-      else if (r.status === 'already_handled') counts.alreadyHeld += 1;
-      else if (r.status === 'undeliverable') counts.undeliverable += 1;
-      else counts.notFound += 1;
-    }
-    return counts;
-  }
-
-  /**
    * Bulk delete prospects (web admin). Fan-out over the hardened single delete — each row
    * keeps its transactional-outbox lead.deleted (mktr-leads-owned rows only; a Lyfe-owned
    * row's app copy is orphaned, the same documented limitation as single delete). One bad
@@ -3004,28 +1685,6 @@ export function makeProspectService(overrides = {}) {
     return counts;
   }
 
-  /**
-   * All ProspectActivity rows for a prospect (oldest-first), with actorUserId explicitly
-   * selected (the getProspect include omits it). Read-only; powers the external lead-timeline
-   * endpoint that feeds the mktr-leads held detail's merged history.
-   */
-  async function getProspectActivities(prospectId, { limit = 200 } = {}) {
-    if (!prospectId) return [];
-    const rows = await m.ProspectActivity.findAll({
-      where: { prospectId },
-      attributes: ['id', 'type', 'description', 'actorUserId', 'metadata', 'createdAt'],
-      order: [['createdAt', 'ASC']],
-      limit: Math.min(parseInt(limit, 10) || 200, 500),
-    });
-    return rows.map((a) => ({
-      id: a.id,
-      type: a.type,
-      description: a.description,
-      actorUserId: a.actorUserId,
-      metadata: a.metadata,
-      createdAt: a.createdAt,
-    }));
-  }
 
   return {
     createProspect,
